+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod convert;
 pub mod delta_e;
 pub mod distinct;
 mod helper;
+pub mod hsv;
 pub mod matrix;
 pub mod named;
 pub mod parser;
@@ -11,6 +12,9 @@ pub mod random;
 mod types;
 
 use std::{fmt, str::FromStr};
+
+// Re-export color space types
+pub use hsv::HSVA;
 
 use colorspace::ColorSpace;
 use convert::{gam_srgb, lin_srgb};
@@ -232,28 +236,7 @@ impl Color {
     /// alpha channel is `1.0`, the simplified `hsv()` format will be used instead.
     pub fn to_hsv_string(&self, format: Format) -> String {
         let hsv = HSVA::from(self);
-        let space = if format == Format::Spaces { " " } else { "" };
-        let (a_prefix, a) = if hsv.alpha == 1.0 {
-            ("", "".to_string())
-        } else {
-            (
-                "a",
-                format!(
-                    ",{space}{alpha}",
-                    alpha = MaxPrecision::wrap(3, hsv.alpha),
-                    space = space
-                ),
-            )
-        };
-        format!(
-            "hsv{a_prefix}({h:.0},{space}{s:.1}%,{space}{v:.1}%{a})",
-            space = space,
-            a_prefix = a_prefix,
-            h = hsv.h,
-            s = 100.0 * hsv.s,
-            v = 100.0 * hsv.v,
-            a = a,
-        )
+        hsv.to_color_string(format)
     }
 
     /// Convert a `Color` to its hue, whiteness, blackness, and alpha values. The hue is given in
@@ -800,24 +783,6 @@ impl From<&HSLA> for Color {
     }
 }
 
-impl From<&HSVA> for Color {
-    fn from(color: &HSVA) -> Self {
-        let lightness = color.v * (1.0 - color.s / 2.0);
-        let saturation = if lightness > 0.0 && lightness < 1.0 {
-            (color.v - lightness) / lightness.min(1.0 - lightness)
-        } else {
-            0.0
-        };
-
-        Color {
-            hue: Hue::from(color.h),
-            saturation: clamp(0.0, 1.0, saturation),
-            lightness: clamp(0.0, 1.0, lightness),
-            alpha: clamp(0.0, 1.0, color.alpha),
-        }
-    }
-}
-
 impl From<&HWBA> for Color {
     fn from(color: &HWBA) -> Self {
         if color.w + color.b >= 1.0 {
@@ -1127,63 +1092,6 @@ impl From<&Color> for HSLA {
 impl fmt::Display for HSLA {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "hsl({h}, {s}, {l})", h = self.h, s = self.s, l = self.l,)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct HSVA {
-    pub h: Scalar,
-    pub s: Scalar,
-    pub v: Scalar,
-    pub alpha: Scalar,
-}
-
-impl ColorSpace for HSVA {
-    fn from_color(c: &Color) -> Self {
-        c.to_hsva()
-    }
-
-    fn into_color(self) -> Color {
-        Color::from_hsva(self.h, self.s, self.v, self.alpha)
-    }
-
-    fn mix(&self, other: &Self, fraction: Fraction) -> Self {
-        // make sure that the hue is preserved when mixing with gray colors
-        let self_hue = if self.s < 0.0001 { other.h } else { self.h };
-        let other_hue = if other.s < 0.0001 { self.h } else { other.h };
-
-        Self {
-            h: interpolate_angle(self_hue, other_hue, fraction),
-            s: interpolate(self.s, other.s, fraction),
-            v: interpolate(self.v, other.v, fraction),
-            alpha: interpolate(self.alpha, other.alpha, fraction),
-        }
-    }
-}
-
-impl From<&Color> for HSVA {
-    fn from(color: &Color) -> Self {
-        let lightness = color.lightness;
-
-        let value = lightness + color.saturation * lightness.min(1.0 - lightness);
-        let saturation = if value > 0.0 {
-            2.0 * (1.0 - lightness / value)
-        } else {
-            0.0
-        };
-
-        HSVA {
-            h: color.hue.value(),
-            s: saturation,
-            v: value,
-            alpha: color.alpha,
-        }
-    }
-}
-
-impl fmt::Display for HSVA {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "hsv({h}, {s}, {v})", h = self.h, s = self.s, v = self.v)
     }
 }
 
@@ -1688,25 +1596,6 @@ mod tests {
         assert_eq!(0xff0000, Color::red().to_u32());
         assert_eq!(0xffffff, Color::white().to_u32());
         assert_eq!(0xf4230f, Color::from_rgb(0xf4, 0x23, 0x0f).to_u32());
-    }
-
-    #[test]
-    fn hsva_conversion() {
-        assert_eq!(
-            Color::from_hsla(0.0, 1.0, 0.5, 0.5),
-            Color::from_hsva(0.0, 1.0, 1.0, 0.5)
-        );
-
-        let roundtrip = |h, s, l| {
-            let color1 = Color::from_hsl(h, s, l);
-            let hsva1 = color1.to_hsva();
-            let color2 = Color::from_hsva(hsva1.h, hsva1.s, hsva1.v, hsva1.alpha);
-            assert_almost_equal(&color1, &color2);
-        };
-
-        for hue in 0..360 {
-            roundtrip(Scalar::from(hue), 0.2, 0.8);
-        }
     }
 
     #[test]
